core/requirements.txt
djangorestframework
django
os
